Move MessageTextIndexSettings and RelatedTermIndexSettings to the files where they are used. Move some imports to the top level. Probably more stuff I made Claude do
