+++ python/ta/test/test_propindex.py
@@ -57,10 +57,8 @@ class SimpleFakeConversation(IConversation):
         # Create storage provider with test settings
         from typeagent.aitools.embeddings import AsyncEmbeddingModel, TEST_MODEL_NAME
         from typeagent.aitools.vectorbase import TextEmbeddingIndexSettings
-        from typeagent.knowpro.importing import (
-            MessageTextIndexSettings,
-            RelatedTermIndexSettings,
-        )
+        from typeagent.knowpro.messageindex import MessageTextIndexSettings
+        from typeagent.knowpro.reltermsindex import RelatedTermIndexSettings
 
         test_model = AsyncEmbeddingModel(model_name=TEST_MODEL_NAME)
         embedding_settings = TextEmbeddingIndexSettings(test_model)
@@ -315,10 +313,8 @@ class FakeConversation[
         # Create storage provider with test settings
         from typeagent.aitools.embeddings import AsyncEmbeddingModel, TEST_MODEL_NAME
         from typeagent.aitools.vectorbase import TextEmbeddingIndexSettings
-        from typeagent.knowpro.importing import (
-            MessageTextIndexSettings,
-            RelatedTermIndexSettings,
-        )
+        from typeagent.knowpro.messageindex import MessageTextIndexSettings
+        from typeagent.knowpro.reltermsindex import RelatedTermIndexSettings
 
         test_model = AsyncEmbeddingModel(model_name=TEST_MODEL_NAME)
         embedding_settings = TextEmbeddingIndexSettings(test_model)
